Edge/requirements.txt
Flask==3.0.3
flask_cors==5.0.0
matplotlib==3.9.2
numpy==2.1.2
paho_mqtt==2.1.0
pymavlink==2.4.41
pyserial==3.5